ref: Move transports out of sentry-core
This gets rid of some feature flags in sentry_core \o/
