+++ sentry-core/Cargo.toml
@@ -17,10 +17,7 @@ edition = "2018"
 all-features = true
 
 [features]
-default = ["with_client_implementation", "with_default_transport", "with_panic", "with_failure", "with_device_info", "with_rust_info"]
-with_reqwest_transport = ["reqwest", "httpdate", "with_client_implementation"]
-with_curl_transport = ["curl", "httpdate", "serde_json", "with_client_implementation"]
-with_default_transport = ["with_reqwest_transport", "with_native_tls"]
+default = ["with_client_implementation", "with_panic", "with_failure", "with_device_info", "with_rust_info"]
 with_client_implementation = ["im", "url", "with_backtrace", "rand"]
 with_backtrace = ["backtrace", "regex"]
 with_panic = ["with_backtrace"]
@@ -33,8 +30,6 @@ with_device_info = ["libc", "hostname", "uname", "with_client_implementation"]
 with_rust_info = ["rustc_version", "with_client_implementation"]
 with_debug_meta = ["findshlibs", "with_client_implementation"]
 with_test_support = []
-with_rustls = ["reqwest/rustls-tls"]
-with_native_tls = ["reqwest/default-tls"]
 
 [dependencies]
 backtrace = { version = "0.3.44", optional = true }
@@ -43,7 +38,6 @@ failure = { version = "0.1.6", optional = true }
 log = { version = "0.4.8", optional = true, features = ["std"] }
 sentry-types = { path = "../sentry-types", version = "0.15.0" }
 env_logger = { version = "0.7.1", optional = true }
-reqwest = { version = "0.10.1", optional = true, features = ["blocking", "json"], default-features = false }
 lazy_static = "1.4.0"
 regex = { version = "1.3.4", optional = true }
 error-chain = { version = "0.12.1", optional = true }
@@ -52,9 +46,6 @@ libc = { version = "0.2.66", optional = true }
 hostname = { version = "0.3.0", optional = true }
 findshlibs = { version = "0.5.0", optional = true }
 rand = { version = "0.7.3", optional = true }
-httpdate = { version = "0.3.2", optional = true }
-curl = { version = "0.4.25", optional = true }
-serde_json = { version = "1.0.48", optional = true }
 
 [target."cfg(not(windows))".dependencies]
 uname = { version = "0.1.1", optional = true }
